+++ src/interface/presenter/components/Layout.tsx
@@ -1,0 +1,25 @@
+import React from 'react'
+import { createStyles, makeStyles, Theme } from '@material-ui/core/styles'
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      flexGrow: 1,
+    },
+  }),
+);
+
+
+type LayoutProps = {
+  children?: React.ReactNode
+}
+
+export const Layout = ({ children }: LayoutProps) => {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      {children}
+    </div>
+  )
+}
